letta_hybrid_poc/pyproject.toml
[tool.poetry]
name = "letta-hybrid-poc"
version = "0.1.0"
description = "Letta + Hybrid Retrieval (BM25 + FAISS) + Cross-Encoder Rerank + Self-RAG POC"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
typer = "^0.9.0"
rich = "^13.7.0"
numpy = "^1.26.2"
scipy = "^1.11.4"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
faiss-cpu = "^1.7.4"
rank-bm25 = "^0.2.2"
sentence-transformers = "^2.2.2"
transformers = "^4.35.2"
torch = "^2.1.1"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
networkx = "^3.2.1"
diskcache = "^5.6.3"
openai = "^1.0.0"
sseclient-py = "^1.8.0"
mem0ai = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
